requirements.txt
Django==2.2.7
Pillow==7.1.2
pyodbc==4.0.30
PyPDF2==1.26.0
pytils==0.3
pytz==2020.1
setuptools==20.7.0
sqlparse==0.3.1
xlrd==1.2.0
